Apply PR corrections - moving control from storage to backend
